requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
cmake=3.25.0=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
einops=0.7.0=pypi_0
filelock=3.9.0=pypi_0
fonttools=4.51.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
huggingface-hub=0.22.2=pypi_0
idna=3.4=pypi_0
jax=0.4.26=pypi_0
jaxlib=0.4.26=pypi_0
jinja2=3.1.2=pypi_0
joypy=0.2.6=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.11.2=pypi_0
lit=15.0.7=pypi_0
mamba-ssm=1.2.0.post1=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.4=pypi_0
ml-dtypes=0.4.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.2.1=pypi_0
ninja=1.11.1.1=pypi_0
numpy=1.26.3=pypi_0
openssl=3.0.13=h7f8727e_0
opt-einsum=3.3.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py310h06a4308_0
pyparsing=3.1.2=pypi_0
python=3.10.14=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-lightning=2.2.2=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2024.4.16=pypi_0
requests=2.28.1=pypi_0
s5-pytorch=0.1.9=pypi_0
safetensors=0.4.3=pypi_0
scipy=1.13.0=pypi_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.19.1=pypi_0
torch=2.0.0+cu118=pypi_0
torchaudio=2.0.1+cu118=pypi_0
torchmetrics=1.3.2=pypi_0
torchvision=0.15.1+cu118=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.40.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=1.26.13=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
yarl=1.9.4=pypi_0
zlib=1.2.13=h5eee18b_0
